refactor: 임시 SNB 구조 리팩
