+++ src/App.tsx
@@ -11,7 +11,7 @@ import { theme } from '@/common/style/theme/theme';
 import { HTTPError } from '@/shared/api/HTTPError';
 import Login from '@/shared/component/Login/Login';
 import ModalContainer from '@/shared/component/Modal/ModalContainer';
-import TempSNB from '@/shared/component/SideNavBar/TempSNB';
+import SNB from '@/shared/component/SideNavBar/LeftSidebar';
 import { HTTP_STATUS_CODE } from '@/shared/constant/api';
 import { PATH } from '@/shared/constant/path';
 import ErrorPage from '@/shared/page/errorPage/ErrorPage';
@@ -49,7 +49,7 @@ const App = () => {
     <ErrorBoundary fallback={ErrorPage} onReset={handleResetError}>
       <Login>
         <ModalContainer />
-        <TempSNB />
+        <SNB />
         <main css={layoutStyle}>
           <Outlet />
         </main>
