+++ src/shared/component/SideNavBar/LeftSidebarItem/LeftSidebarMenuItem.style.ts
@@ -1,0 +1,23 @@
+import { css } from '@emotion/react';
+
+export const leftSidebarMenuItemStyle = css({
+  display: 'flex',
+  alignItems: 'center',
+
+  cursor: 'pointer',
+});
+
+export const clubInfoStyle = css({
+  padding: '0.4rem',
+
+  borderRadius: '10px',
+});
+
+export const clubLogoStyle = css({
+  width: '4rem',
+  height: '4rem',
+
+  borderRadius: '10px',
+
+  objectFit: 'cover',
+});
