+++ src/shared/component/SideNavBar/LeftSidebarItem/PageIndicatorStick/PageIndicatorStick.style.ts
