+++ src/shared/component/SideNavBar/LeftSidebarItem/PageIndicatorStick/PageIndicatorStick.tsx
@@ -1,4 +1,4 @@
-import { PageIndicatorStickStyle } from '@/shared/component/LeftSidebar/LeftSidebarItem/PageIndicatorStick/PageIndicatorStick.style';
+import { PageIndicatorStickStyle } from '@/shared/component/SideNavBar/LeftSidebarItem/PageIndicatorStick/PageIndicatorStick.style';
 
 interface PageIndicatorStickProps {
   isClicked: boolean;
