--- src/shared/component/SideNavBar/TempSNB.tsx
@@ -1,7 +1,0 @@
-import { tempSNBStyle } from '@/shared/component/SideNavBar/TempSnb.style';
-
-const TempSNB = () => {
-  return <aside css={tempSNBStyle} />;
-};
-
-export default TempSNB;
