--- src/shared/component/SideNavBar/TempSnb.style.ts
@@ -1,10 +1,0 @@
-import { css } from '@emotion/react';
-
-export const tempSNBStyle = css({
-  position: 'fixed',
-
-  width: '7.6rem',
-  height: '100vh',
-
-  backgroundColor: '#E3E8FF',
-});
