--- src/story/shared/LeftSidebarItem.stories.tsx
@@ -1,32 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import LeftSidebarMenuItem from '@/shared/component/LeftSidebar/LeftSidebarItem/LeftSidebarMenuItem';
-
-const meta = {
-  title: 'Shared/LeftSidebarItem',
-  component: LeftSidebarMenuItem,
-  parameters: {
-    layout: 'centered',
-  },
-  args: {
-    children: 'Team 1',
-    isExpanded: true,
-    isClicked: true,
-    logoUrl: 'src/common/asset/svg/ic_add_white.svg',
-    onClick: () => {},
-  },
-  argTypes: {
-    children: {
-      control: {
-        type: 'text',
-      },
-    },
-  },
-} satisfies Meta<typeof LeftSidebarMenuItem>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  args: {},
-};
